+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Display cluster total memory size and number of GPUs in dashboard stats.
   - Display number of GPUs per node in resources page and node details
     page.
+  - Display number of allocated GPU and allocation percentage in node details
+    page.
 - agent: Add total quantity of memory and number of gpus in cluster stats
   response.
 - conf: Select `gres` and `gres_used` attributes by default in `slurmrestd`